moving gson support tests to dedicated class
