+++ nakadi-java-client/src/test/java/nakadi/EventResourceRealTest.java
@@ -446,93 +446,6 @@ public class EventResourceRealTest {
     assertNull(options.getValue().scope());
   }
 
-  @Test
-  public void serdesDomain() {
-    EventResourceReal eventResource = new EventResourceReal(null, new GsonSupport(), new CompressionSupportGzip() );
-
-    EventThing et = new EventThing("a", "b");
-    EventRecord<EventThing> er = new EventRecord<>("topic", et);
-
-    Object o = eventResource.mapEventRecordToSerdes(er);
-
-    assertEquals(et, o);
-  }
-
-  @Test
-  public void serdesUndefinedEventMapped() {
-    EventResourceReal eventResource = new EventResourceReal(null, new GsonSupport(), new CompressionSupportGzip() );
-
-    Map<String, Object> uemap = Maps.newHashMap();
-    uemap.put("a", "1");
-    UndefinedEventMapped<Map<String, Object>> ue =
-        new UndefinedEventMapped<Map<String, Object>>().data(uemap);
-    EventRecord<UndefinedEventMapped> er = new EventRecord<>("topic", ue);
-    Map<String, Object> outmap = (Map<String, Object>) eventResource.mapEventRecordToSerdes(er);
-    assertTrue(outmap.size() == 1);
-    assertTrue(outmap.containsKey("a"));
-    assertEquals("1", outmap.get("a"));
-  }
-
-  @Test
-  public void serdesBusinessEventMapped() {
-    EventResourceReal eventResource = new EventResourceReal(null, new GsonSupport(), new CompressionSupportGzip());
-
-    BusinessEventMapped<Map<String, Object>> be = new BusinessEventMapped<>();
-    EventMetadata em = new EventMetadata();
-    em.eid("eid1").withFlowId().withOccurredAt();
-    Map<String, Object> uemap = Maps.newHashMap();
-    uemap.put("a", 1);
-    uemap.put("b", 22.0001);
-    uemap.put("c", "40.0001");
-    uemap.put("d", "c22.0001");
-    be.data(uemap);
-    be.metadata(em);
-
-    EventRecord<BusinessEventMapped> er = new EventRecord<>("topic", be);
-    JsonObject outmap = (JsonObject) eventResource.mapEventRecordToSerdes(er);
-    assertTrue(outmap.size() == 5);
-    assertTrue(outmap.get("metadata") != null);
-    assertTrue(outmap.get("a") != null);
-    assertEquals(1, outmap.get("a").getAsInt());
-    assertEquals(22.0001, outmap.get("b").getAsDouble(), 0.0d);
-    assertEquals("40.0001", outmap.get("c").getAsString());
-    assertEquals("c22.0001", outmap.get("d").getAsString());
-
-    final JsonElement metadata = outmap.get("metadata");
-    assertEquals(em.eid(), metadata.getAsJsonObject().get("eid").getAsString());
-    assertEquals(em.flowId(), metadata.getAsJsonObject().get("flow_id").getAsString());
-    assertEquals(
-        em.occurredAt(),
-        new OffsetDateTimeSerdes().toOffsetDateTime(
-            metadata.getAsJsonObject().get("occurred_at").getAsString()));
-  }
-
-  static class EventThing implements Event {
-    final String a;
-    final String b;
-
-    public EventThing(String a, String b) {
-      this.a = a;
-      this.b = b;
-    }
-
-    @Override public Object data() {
-      return null;
-    }
-
-    @Override public int hashCode() {
-      return Objects.hash(a, b);
-    }
-
-    @Override public boolean equals(Object o) {
-      if (this == o) return true;
-      if (o == null || getClass() != o.getClass()) return false;
-      EventThing that = (EventThing) o;
-      return Objects.equals(a, that.a) &&
-          Objects.equals(b, that.b);
-    }
-  }
-
   static class UndefinedPayload {
     String id;
     String foo;
